ProcessGridLayer added (aurusov/raox#328)
