+++ ru.bmstu.rk9.rao.ui/src/ru/bmstu/rk9/rao/ui/process/model/ModelFigure.java
@@ -1,22 +1,15 @@
 package ru.bmstu.rk9.rao.ui.process.model;
 
 import org.eclipse.draw2d.Graphics;
+import org.eclipse.draw2d.Layer;
 import org.eclipse.draw2d.XYLayout;
-import org.eclipse.draw2d.geometry.Rectangle;
-import org.eclipse.gef.editparts.GridLayer;
 import org.eclipse.swt.SWT;
 
-public class ModelFigure extends GridLayer {
-
-	private boolean showGrid;
+public class ModelFigure extends Layer {
 
 	public ModelFigure() {
 		setLayoutManager(new XYLayout());
 		setOpaque(false);
-	}
-
-	public final void setShowGrid(boolean showGrid) {
-		this.showGrid = showGrid;
 	}
 
 	@Override
@@ -25,18 +18,5 @@ public class ModelFigure extends GridLayer {
 		graphics.setAntialias(SWT.ON);
 		graphics.setTextAntialias(SWT.ON);
 		super.paint(graphics);
-	}
-
-	@Override
-	protected void paintGrid(Graphics g) {
-		if (!showGrid)
-			return;
-
-		Rectangle clip = g.getClip(Rectangle.SINGLETON);
-		for (int x = clip.x - clip.x % gridX; x <= clip.x + clip.width; x += gridX) {
-			for (int y = clip.y - clip.y % gridY; y <= clip.y + clip.height; y += gridY) {
-				g.drawPoint(x, y);
-			}
-		}
 	}
 }
